# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"What to Wear — RAG styling engine (test API)","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Health Get"}}}}}}},"/suggest":{"post":{"summary":"Suggest Endpoint","description":"Supersedes /recommend (contracts/suggest.md). Same auth model — the\nrequester's identity always comes from the verified JWT `sub`, never the\nbody. SSE: an `outfit` event per ranked outfit, then a `done` event\ncarrying the full response shape (a client that only reads `done` gets\nthe exact non-streaming payload). `response_model` is for OpenAPI docs\nonly — returning a Response subclass directly makes FastAPI skip\nresponse_model serialization, but it's still what generates the\nSuggestResult/ScoredOutfit/DimensionScore schemas the frontend needs\n(T036b) since a raw StreamingResponse alone wouldn't.","operationId":"suggest_endpoint_suggest_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuggestRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuggestResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/wardrobe/items":{"get":{"summary":"List Wardrobe Items","operationId":"list_wardrobe_items_wardrobe_items_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WardrobeItem"},"type":"array","title":"Response List Wardrobe Items Wardrobe Items Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"summary":"Add Wardrobe Item","operationId":"add_wardrobe_item_wardrobe_items_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddWardrobeItemRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WardrobeItem"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/catalog/items":{"get":{"summary":"List Catalog Items","operationId":"list_catalog_items_catalog_items_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WardrobeItem"},"type":"array","title":"Response List Catalog Items Catalog Items Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/wardrobe/items/bulk":{"post":{"summary":"Add Wardrobe Items Bulk","operationId":"add_wardrobe_items_bulk_wardrobe_items_bulk_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddWardrobeItemsBulkRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/WardrobeItem"},"type":"array","title":"Response Add Wardrobe Items Bulk Wardrobe Items Bulk Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/wardrobe/items/{item_id}":{"patch":{"summary":"Update Wardrobe Item","operationId":"update_wardrobe_item_wardrobe_items__item_id__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Item Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WardrobeItemPatch"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WardrobeItem"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Wardrobe Item","operationId":"delete_wardrobe_item_wardrobe_items__item_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Item Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/wardrobe/items/extract":{"post":{"summary":"Extract Wardrobe Item","description":"Draft extraction only — nothing is persisted to wardrobe_items here\n(US2). A photo that can't be confidently interpreted is still a 200 with\nextraction_ok=False, never a 5xx (FR-006).","operationId":"extract_wardrobe_item_wardrobe_items_extract_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_extract_wardrobe_item_wardrobe_items_extract_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoExtractionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/wardrobe/items/upload":{"post":{"summary":"Upload Wardrobe Item","description":"Persists a wardrobe item from user-confirmed (possibly corrected)\nattributes -- source='upload', parallel to, not replacing, the\ncatalog-based POST /wardrobe/items (US2).","operationId":"upload_wardrobe_item_wardrobe_items_upload_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWardrobeItemFromUploadRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WardrobeItem"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/preferences/feedback":{"post":{"summary":"Submit Feedback","description":"Records or replaces a reaction to a specific outfit (US1). item_ids\nare resolved against the caller's own wardrobe -- an id that doesn't\nexist or belongs to someone else is a 404, never silently accepted.","operationId":"submit_feedback_preferences_feedback_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitFeedbackRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuggestionFeedback"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/preferences":{"get":{"summary":"Get Preferences","description":"A plain-language summary of what's been learned (US3). has_feedback\ndistinguishes \"no feedback at all\" from \"feedback exists, no signal has\ncrossed threshold yet\" -- both have signals=[], only the former is False.","operationId":"get_preferences_preferences_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PreferenceProfile"}}}}},"security":[{"HTTPBearer":[]}]},"delete":{"summary":"Clear Preferences","description":"Clears the entire derived profile in one action (US4) -- dismisses\nevery signal currently present, reusing the exact same per-signal\nmechanism as remove_preference_signal rather than a separate code path\n(research.md #3).","operationId":"clear_preferences_preferences_delete","responses":{"204":{"description":"Successful Response"}},"security":[{"HTTPBearer":[]}]}},"/preferences/signals/{signal_key}":{"delete":{"summary":"Remove Preference Signal","description":"Removes one derived signal without affecting the rest of the profile\n(US4). Idempotent -- dismissing an already-absent signal is a no-op,\nnot a 404 (contracts/preferences.md).","operationId":"remove_preference_signal_preferences_signals__signal_key__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"signal_key","in":"path","required":true,"schema":{"type":"string","title":"Signal Key"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AddWardrobeItemRequest":{"properties":{"catalog_item_id":{"type":"string","format":"uuid","title":"Catalog Item Id"}},"type":"object","required":["catalog_item_id"],"title":"AddWardrobeItemRequest"},"AddWardrobeItemsBulkRequest":{"properties":{"catalog_item_ids":{"items":{"type":"string","format":"uuid"},"type":"array","title":"Catalog Item Ids"}},"type":"object","required":["catalog_item_ids"],"title":"AddWardrobeItemsBulkRequest"},"Body_extract_wardrobe_item_wardrobe_items_extract_post":{"properties":{"photo":{"type":"string","contentMediaType":"application/octet-stream","title":"Photo"}},"type":"object","required":["photo"],"title":"Body_extract_wardrobe_item_wardrobe_items_extract_post"},"CitedSource":{"properties":{"rule_id":{"type":"string","title":"Rule Id"},"source":{"type":"string","title":"Source"},"url":{"type":"string","title":"Url"},"layer":{"type":"string","enum":["L1","L2","L3","L4"],"title":"Layer"}},"type":"object","required":["rule_id","source","url","layer"],"title":"CitedSource"},"Context":{"properties":{"occasion":{"type":"string","title":"Occasion"},"formality":{"type":"string","enum":["casual","smart_casual","business_casual","semi_formal","formal","black_tie"],"title":"Formality"},"mood":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mood"},"temp_c":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Temp C"},"condition":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Condition"},"temp_band":{"anyOf":[{"type":"string","enum":["freezing","cold","cool","mild","warm","hot"]},{"type":"null"}],"title":"Temp Band"},"season":{"anyOf":[{"type":"string","enum":["spring","summer","autumn","winter"]},{"type":"null"}],"title":"Season"},"wardrobe":{"items":{"$ref":"#/components/schemas/WardrobeItem"},"type":"array","title":"Wardrobe"},"user_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Id"}},"type":"object","required":["occasion","formality"],"title":"Context","description":"Normalized request context (pipeline stage 1 output)."},"CreateWardrobeItemFromUploadRequest":{"properties":{"photo_path":{"type":"string","title":"Photo Path"},"category":{"type":"string","title":"Category"},"colors":{"items":{"type":"string"},"type":"array","minItems":1,"title":"Colors"},"formality":{"type":"string","enum":["casual","smart_casual","business_casual","semi_formal","formal","black_tie"],"title":"Formality"},"warmth":{"type":"integer","maximum":5.0,"minimum":0.0,"title":"Warmth"},"season":{"items":{"type":"string","enum":["spring","summer","autumn","winter"]},"type":"array","minItems":1,"title":"Season"},"fabric":{"type":"string","title":"Fabric"},"pattern":{"type":"string","title":"Pattern"},"fit":{"type":"string","title":"Fit"}},"type":"object","required":["photo_path","category","colors","formality","warmth","season","fabric","pattern","fit"],"title":"CreateWardrobeItemFromUploadRequest","description":"Body of POST /wardrobe/items/upload — the user-confirmed (possibly\ncorrected) attributes. fabric/pattern/fit are required HERE ONLY: SC-003\nrequires 100% of items saved through the photo flow to have every\nattribute populated, none blank. WardrobeItem/WardrobeItemPatch (the\ncorrection path) keep all three optional, unchanged."},"DimensionScore":{"properties":{"dimension":{"type":"string","enum":["color_harmony","formality_coherence","weather_fitness","silhouette_balance"],"title":"Dimension"},"value":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Value"},"reason":{"type":"string","title":"Reason"}},"type":"object","required":["dimension","value","reason"],"title":"DimensionScore","description":"One deterministic scorer's output for one outfit (data-model.md)."},"ExtractedAttributes":{"properties":{"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"colors":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Colors"},"fabric":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fabric"},"warmth":{"anyOf":[{"type":"integer","maximum":5.0,"minimum":0.0},{"type":"null"}],"title":"Warmth"},"formality":{"anyOf":[{"type":"string","enum":["casual","smart_casual","business_casual","semi_formal","formal","black_tie"]},{"type":"null"}],"title":"Formality"},"season":{"anyOf":[{"items":{"type":"string","enum":["spring","summer","autumn","winter"]},"type":"array"},{"type":"null"}],"title":"Season"},"pattern":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pattern"},"fit":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fit"}},"type":"object","title":"ExtractedAttributes","description":"Draft output of one VLM extraction call over a single item photo.\nEvery field optional — extraction failing on any/all of them must not\nblock adding the item (FR-006); the user fills in whatever's missing."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PhotoExtractionResponse":{"properties":{"photo_path":{"type":"string","title":"Photo Path"},"extracted":{"$ref":"#/components/schemas/ExtractedAttributes"},"extraction_ok":{"type":"boolean","title":"Extraction Ok"}},"type":"object","required":["photo_path","extracted","extraction_ok"],"title":"PhotoExtractionResponse","description":"What POST /wardrobe/items/extract returns — an unsaved draft."},"PreferenceProfile":{"properties":{"has_feedback":{"type":"boolean","title":"Has Feedback"},"signals":{"items":{"$ref":"#/components/schemas/PreferenceSignal"},"type":"array","title":"Signals"}},"type":"object","required":["has_feedback"],"title":"PreferenceProfile","description":"What GET /preferences returns. has_feedback distinguishes \"no\nfeedback at all\" (FR-008's empty state) from \"feedback exists but no\nsignal has crossed threshold yet\" (also signals=[], but has_feedback is\nTrue)."},"PreferenceSignal":{"properties":{"key":{"type":"string","title":"Key"},"summary":{"type":"string","title":"Summary"}},"type":"object","required":["key","summary"],"title":"PreferenceSignal"},"Rationale":{"properties":{"text":{"type":"string","title":"Text"},"cites":{"items":{"type":"string"},"type":"array","title":"Cites"}},"type":"object","required":["text"],"title":"Rationale"},"ScoredOutfit":{"properties":{"items":{"items":{"type":"string"},"type":"array","title":"Items"},"rationale":{"items":{"$ref":"#/components/schemas/Rationale"},"type":"array","title":"Rationale"},"scores":{"items":{"$ref":"#/components/schemas/DimensionScore"},"type":"array","title":"Scores"},"rank_score":{"type":"number","title":"Rank Score"}},"type":"object","required":["items","rationale","scores","rank_score"],"title":"ScoredOutfit","description":"An `Outfit` extended with per-dimension scores and a combined rank\nscore (data-model.md). `scores` always carries exactly one entry per\n`SCORE_DIMENSIONS` value (FR-008)."},"SubmitFeedbackRequest":{"properties":{"verdict":{"type":"string","enum":["liked","rejected"],"title":"Verdict"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"},"item_ids":{"items":{"type":"string"},"type":"array","minItems":1,"title":"Item Ids"}},"type":"object","required":["verdict","item_ids"],"title":"SubmitFeedbackRequest","description":"Body of POST /preferences/feedback. item_ids identify the reacted-to\noutfit -- resolved server-side against the caller's own wardrobe_items,\nnever trusted as-is (see crud.record_feedback)."},"SuggestRequest":{"properties":{"occasion":{"type":"string","title":"Occasion"},"mood":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mood"},"formality":{"anyOf":[{"type":"string","enum":["casual","smart_casual","business_casual","semi_formal","formal","black_tie"]},{"type":"null"}],"title":"Formality"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location"},"temp_c":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Temp C"},"strategy":{"type":"string","title":"Strategy","default":"advanced"},"thread_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thread Id"}},"type":"object","required":["occasion"],"title":"SuggestRequest","description":"Body of POST /suggest (contracts/suggest.md). No `user_id` field —\nsame fix as RecommendRequest post-Phase-1: identity always comes from\nthe verified JWT `sub` (FR-001)."},"SuggestResult":{"properties":{"outfits":{"items":{"$ref":"#/components/schemas/ScoredOutfit"},"type":"array","title":"Outfits"},"sources":{"items":{"$ref":"#/components/schemas/CitedSource"},"type":"array","title":"Sources"},"context":{"anyOf":[{"$ref":"#/components/schemas/Context"},{"type":"null"}]}},"type":"object","required":["outfits"],"title":"SuggestResult","description":"What POST /suggest's `result` field carries (contracts/suggest.md) —\nOutfitResult's shape, but with ScoredOutfit entries. A separate type\nfrom OutfitResult (not a field-type change on it) since OutfitResult is\nkept as `pipeline.cite.build_result`'s internal return shape (see its\ndocstring above) rather than merged away."},"SuggestionFeedback":{"properties":{"id":{"type":"string","title":"Id"},"verdict":{"type":"string","enum":["liked","rejected"],"title":"Verdict"},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"},"item_ids":{"items":{"type":"string"},"type":"array","title":"Item Ids"},"created_at":{"type":"string","title":"Created At"}},"type":"object","required":["id","verdict","item_ids","created_at"],"title":"SuggestionFeedback","description":"What POST /preferences/feedback returns."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WardrobeItem":{"properties":{"id":{"type":"string","title":"Id"},"category":{"type":"string","title":"Category"},"colors":{"items":{"type":"string"},"type":"array","title":"Colors"},"formality":{"type":"string","enum":["casual","smart_casual","business_casual","semi_formal","formal","black_tie"],"title":"Formality"},"warmth":{"type":"integer","maximum":5.0,"minimum":0.0,"title":"Warmth"},"season":{"items":{"type":"string","enum":["spring","summer","autumn","winter"]},"type":"array","title":"Season"},"fabric":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fabric"},"source":{"anyOf":[{"type":"string","enum":["catalog","upload"]},{"type":"null"}],"title":"Source"},"pattern":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pattern"},"fit":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fit"}},"type":"object","required":["id","category","formality","warmth"],"title":"WardrobeItem","description":"A single owned garment. Wardrobe capture is out of scope this phase;\nthis is the shape the capture flow must eventually emit.\n\n`colors` are hex — the source of truth (see colors.py). Human-readable\nnames are derived on demand via colors.nearest_names(), never stored, so\nname and hex can never drift out of sync."},"WardrobeItemPatch":{"properties":{"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"colors":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Colors"},"formality":{"anyOf":[{"type":"string","enum":["casual","smart_casual","business_casual","semi_formal","formal","black_tie"]},{"type":"null"}],"title":"Formality"},"warmth":{"anyOf":[{"type":"integer","maximum":5.0,"minimum":0.0},{"type":"null"}],"title":"Warmth"},"season":{"anyOf":[{"items":{"type":"string","enum":["spring","summer","autumn","winter"]},"type":"array"},{"type":"null"}],"title":"Season"},"fabric":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fabric"},"pattern":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pattern"},"fit":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fit"}},"type":"object","title":"WardrobeItemPatch","description":"A partial correction to an owned WardrobeItem (US3). Every field is\noptional -- only fields present in the request are applied, matching\nPATCH semantics. Reuses the same field-level validation as WardrobeItem\nso an invalid value is rejected with a 422 before it ever reaches the DB\n(FR-007), while `category` stays open-ended (its slot/bucket is derived\non read via categories.group_of(), never itself validated here)."}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}